refactor(api): consolidate remaining route registration wiring
